+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "skyvern"
-version = "0.1.88"
+version = "0.1.89"
 description = ""
 authors = ["Skyvern AI <info@skyvern.com>"]
 readme = "README.md"
