+++ skyvern/cli/quickstart.py
@@ -2,7 +2,6 @@
 
 import asyncio
 import subprocess
-import sys
 
 import typer
 from rich.panel import Panel
@@ -11,6 +10,7 @@ from rich.progress import Progress, SpinnerColumn, TextColumn
 # Import console after skyvern.cli to ensure proper initialization
 from skyvern.cli.console import console
 from skyvern.cli.init_command import init  # init is used directly
+from skyvern.cli.utils import start_services
 
 quickstart_app = typer.Typer(help="Quickstart command to set up and run Skyvern with one command.")
 
@@ -27,40 +27,6 @@ def check_docker() -> bool:
         return result.returncode == 0
     except (FileNotFoundError, subprocess.SubprocessError):
         return False
-
-
-async def start_services(server_only: bool = False) -> None:
-    """Start Skyvern services in the background.
-
-    Args:
-        server_only: If True, only start the server, not the UI.
-    """
-    try:
-        # Start server in the background
-        server_process = await asyncio.create_subprocess_exec(
-            sys.executable, "-m", "skyvern.cli.commands", "run", "server"
-        )
-
-        # Give server a moment to start
-        await asyncio.sleep(2)
-
-        if not server_only:
-            # Start UI in the background
-            ui_process = await asyncio.create_subprocess_exec(sys.executable, "-m", "skyvern.cli.commands", "run", "ui")
-
-        console.print("\n🎉 [bold green]Skyvern is now running![/bold green]")
-        console.print("🌐 [bold]Access the UI at:[/bold] [cyan]http://localhost:8080[/cyan]")
-        console.print("🔑 [bold]Your API key is in your .env file as SKYVERN_API_KEY[/bold]")
-
-        # Wait for processes to complete (they won't unless killed)
-        if not server_only:
-            await asyncio.gather(server_process.wait(), ui_process.wait())
-        else:
-            await server_process.wait()
-
-    except Exception as e:
-        console.print(f"[bold red]Error starting services: {str(e)}[/bold red]")
-        raise typer.Exit(1)
 
 
 @quickstart_app.callback(invoke_without_command=True)
